Friday P5 Mr Dick
Useful notes. This isnt everything on capital structure, as we only touched on it for our mocks, it is mainly taught in y13. But here are the basics of everything.
If you have any questions email me, and ill get back to you when i can, as i have some more in-depth notes.
Also Monday, in B6 for lesson.

Signed-off-by: JEames06 <117163576+JEames06@users.noreply.github.com>